Backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
pydantic
